Added a project for Visual Studio 2019.
- Updated the README with instructions for the new and old project, as well as just a general update on the compilation instructions.
- Fixed some parts of the code that were causing errors on VS 2019.
- The .rc file included in this commit is the one VS 2019 uses. The Pikifen.rc file found in /Source/source is only used by VS 2013.

+++ Source/source/mobs/mob.cpp
@@ -409,7 +409,7 @@ void mob::arachnorb_plan_logic(const unsigned char goal) {
     
     amount_to_move = min(amount_to_move, max_step_distance);
     amount_to_turn =
-        sign(amount_to_turn) * min(fabs(amount_to_turn), max_turn_angle);
+        sign(amount_to_turn) * min((double) fabs(amount_to_turn), (double) max_turn_angle);
         
     point destination_pos = pos;
     float destination_angle = angle + amount_to_turn;
